3. Add animation rhythm over time
To improve the code's visual appeal, I drew inspiration from comics and incorporated effects that utilize the concept of time. Due to the addition of speed and color changes, I established new definitions to accommodate the code changes, utilizing millis() to control the segmentation of time and lerp() and map() to manage the visual element changes. Also, I changed the speed of the pointer movement so that the movement of the pointer is also involved in the overall animation.